requirements.txt
dbt-core
dbt-snowflake
airflow-apache